fix bug cannot add subject Change CreateClass Modal Delete CreateClass Component
